Move HASH() and ORA_HASH() functions into own file
