+++ h2/src/main/org/h2/expression/function/CryptFunction.java
@@ -64,10 +64,9 @@ public final class CryptFunction extends OperationN implements NamedExpression {
         cipher.setKey(getPaddedArrayCopy(v2.getBytesNoCopy(), cipher.getKeyLength()));
         byte[] newData = getPaddedArrayCopy(v3.getBytesNoCopy(), BlockCipher.ALIGN);
         switch (function) {
-        case ENCRYPT: {
+        case ENCRYPT:
             cipher.encrypt(newData, 0, newData.length);
             break;
-        }
         case DECRYPT:
             cipher.decrypt(newData, 0, newData.length);
             break;
